backend/backend_app/requirements.txt
fastapi
uvicorn
asterisk-ami
telnetlib3
pystrix